Hook up retry handler for unexpected exception.
This is divergent from zipline, where backtest is unlikely
fails or code error should be reported more quickly.
In the case of live trading, there are many chances that
program encounters unexpected external issues. As the
scheduler of algorithm, executor should not give up easily.

This is to catch general Exceptions.  KeyboardInterrupt and more
critical error still exits.