SESSION_END should be yielded once per day
the SESSION_END and MINUTE_END are used for performance logging in the
original package: zipline and zipline-live.
we currently do not log performance but we do yeild those events.
without this fix the events will be yielded time and time again the entire lasr minute
of the day (session_close). which is of course not a desired action.