removing __setattr__ protection. it's not need in this kind of project
__setattr__ protects the module from overriding methods. it is useful for
hosted packages. when we don't want the user to override context methods.
it's done by quantopian because they are running their solution in the cloud.
but this project is executed by every user in their local host. if they want
to override something they should be able to do so
plus, by removing this we allow the smoke tests to be available again. they were
broken because we could overrider the pipeline output method. this will fix this
issue.