I am basically reverting this commit 231516
https://github.com/alpacahq/pylivetrader/commit/231516323c90e6e59fe865cf2d2b3654d227e40c
so I wanted to allow the user to be able to set attribute as they desire, BUT
it doesn't protect them from a doing setattr by mistake and overriding methods.
example given: https://github.com/alpacahq/pylivetrader/issues/76
in that example a context.symbol was overridden by mistake abd it caused a broken
environment. so by bringing this back we protect the user from making such mistakes.